Move Notification_Ui.py to ui/NotificationWindow.py
