+++ mTwit/ui/MainWindow.py
@@ -24,7 +24,7 @@ from PyQt5.QtWidgets import (
 from system_hotkey import SystemHotkey
 
 from mTwit.ui.ui_base import Win32Window, HoverButton, QuitButton
-from mTwit.Notification_Ui import *
+from .NotificationWindow import NotificationWindow, NotificationMode
 from .AuthWindow import AuthWindow
 
 class MainWindow(QMainWindow):
