+++ mTwit/ui/__init__.py
@@ -1,2 +1,4 @@
 from . import ui_base
 from .MainWindow import MainWindow
+from .AuthWindow import AuthWindow
+from .NotificationWindow import NotificationWindow
